refactor TmClient into its own module
