+++ src/lib.rs
@@ -5,6 +5,7 @@ use crate::report::Error;
 pub mod config;
 pub mod event_sub;
 pub mod report;
+pub mod tm_client;
 
 pub fn run(_cfg: config::Config) -> Result<(), Error> {
     unimplemented!()
